ai_server/requirements.txt
absl-py==2.2.2
acres==0.3.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
astunparse==1.6.3
certifi==2025.1.31
charset-normalizer==3.4.1
ci-info==0.3.0
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
configobj==5.0.9
configparser==7.2.0
distro==1.9.0
Django==4.2.20
django-cors-headers==4.3.1
djangorestframework==3.16.0
dotenv==0.9.9
etelemetry==0.3.1
exceptiongroup==1.2.2
filelock==3.18.0
flatbuffers==25.2.10
gast==0.6.0
google-pasta==0.2.0
griffe==1.7.3
grpcio==1.71.0
gym==0.26.2
gym-notices==0.0.8
h11==0.14.0
h5py==3.13.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
importlib_resources==6.5.2
isodate==0.6.1
jiter==0.9.0
keras==3.9.2
libclang==18.1.1
looseversion==1.3.0
lxml==5.3.2
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mcp==1.6.0
mdurl==0.1.2
ml_dtypes==0.5.1
mysqlclient==2.2.7
namex==0.0.9
networkx==3.2.1
nibabel==5.3.2
nipype==1.10.0
numpy==2.0.2
openai==1.72.0
openai-agents==0.0.12
opt_einsum==3.4.0
optree==0.15.0
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
protobuf==5.29.4
prov==2.0.1
puremagic==1.28
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
pydot==3.0.4
Pygments==2.19.1
PyMuPDF==1.25.5
PyMySQL==1.1.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pyxnat==1.6.3
rdflib==6.3.2
requests==2.32.3
rich==14.0.0
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
scipy==1.13.1
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
sse-starlette==2.2.1
starlette==0.46.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.0.1
tqdm==4.67.1
traits==7.0.2
types-requests==2.32.0.20250328
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.21.0
